Fix VariableWriter and move test to test_variable_writer.py
